Updates to work better when adding new books
The sources file will no longer immediately update its modified date when loading a library from a json file. This allows the library to better recognize when the sources file has been updated.

Books will now check if their folder exists and has pages content in it when determining if the book has an update available.

The library will now check for existing books by their file path instead of their name. This allows the library to update a book's name if it has changed.